+++ sdk/src/main/java/com/hedera/hashgraph/sdk/TransactionResponse.java
@@ -6,6 +6,7 @@ import org.bouncycastle.util.encoders.Hex;
 
 import javax.annotation.Nullable;
 import java.util.Collections;
+import java.util.concurrent.TimeoutException;
 
 public final class TransactionResponse implements WithGetReceipt, WithGetRecord {
     public final AccountId nodeId;
@@ -29,11 +30,17 @@ public final class TransactionResponse implements WithGetReceipt, WithGetRecord 
         this.scheduledTransactionId = scheduledTransactionId;
     }
 
-    public TransactionReceipt getReceiptSync(Client client) throws Exception {
-        return new TransactionReceiptQuery()
-            .setTransactionId(transactionId)
-            .setNodeAccountIds(Collections.singletonList(nodeId))
-            .execute(client);
+    public TransactionReceipt getReceipt(Client client) throws TimeoutException, PrecheckStatusException, ReceiptStatusException {
+        var receipt = new TransactionReceiptQuery()
+                .setTransactionId(transactionId)
+                .setNodeAccountIds(Collections.singletonList(nodeId))
+                .execute(client);
+
+        if (receipt.status != Status.SUCCESS) {
+            throw new ReceiptStatusException(transactionId, receipt);
+        }
+
+        return receipt;
     }
 
     @Override
@@ -44,8 +51,8 @@ public final class TransactionResponse implements WithGetReceipt, WithGetRecord 
             .executeAsync(client);
     }
 
-    public TransactionRecord getRecordSync(Client client) throws Exception {
-        getReceiptSync(client);
+    public TransactionRecord getRecord(Client client) throws TimeoutException, PrecheckStatusException, ReceiptStatusException {
+        getReceipt(client);
 
         return new TransactionRecordQuery()
             .setTransactionId(transactionId)